Refactor node shutdown manager
