+++ pkg/kubelet/nodeshutdown/nodeshutdown_manager_linux.go
@@ -21,10 +21,8 @@ limitations under the License.
 package nodeshutdown
 
 import (
-	"context"
 	"fmt"
 	"path/filepath"
-	"sort"
 	"sync"
 	"time"
 
@@ -32,23 +30,16 @@ import (
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
 	"k8s.io/client-go/tools/record"
 	"k8s.io/klog/v2"
-	podutil "k8s.io/kubernetes/pkg/api/v1/pod"
-	"k8s.io/kubernetes/pkg/apis/scheduling"
 	"k8s.io/kubernetes/pkg/features"
-	kubeletconfig "k8s.io/kubernetes/pkg/kubelet/apis/config"
 	kubeletevents "k8s.io/kubernetes/pkg/kubelet/events"
 	"k8s.io/kubernetes/pkg/kubelet/eviction"
 	"k8s.io/kubernetes/pkg/kubelet/lifecycle"
 	"k8s.io/kubernetes/pkg/kubelet/metrics"
 	"k8s.io/kubernetes/pkg/kubelet/nodeshutdown/systemd"
 	"k8s.io/kubernetes/pkg/kubelet/prober"
-	"k8s.io/kubernetes/pkg/kubelet/volumemanager"
-	"k8s.io/utils/clock"
 )
 
 const (
-	nodeShutdownReason             = "Terminated"
-	nodeShutdownMessage            = "Pod was terminated in response to imminent node shutdown."
 	nodeShutdownNotAdmittedReason  = "NodeShutdown"
 	nodeShutdownNotAdmittedMessage = "Pod was rejected as the node is shutting down."
 	dbusReconnectPeriod            = 1 * time.Second
@@ -75,12 +66,7 @@ type managerImpl struct {
 	nodeRef      *v1.ObjectReference
 	probeManager prober.Manager
 
-	volumeManager volumemanager.VolumeManager
-
-	shutdownGracePeriodByPodPriority []kubeletconfig.ShutdownGracePeriodByPodPriority
-
 	getPods        eviction.ActivePodsFunc
-	killPodFunc    eviction.KillPodFunc
 	syncNodeStatus func()
 
 	dbusCon     dbusInhibiter
@@ -88,53 +74,36 @@ type managerImpl struct {
 
 	nodeShuttingDownMutex sync.Mutex
 	nodeShuttingDownNow   bool
-
-	clock clock.Clock
+	podManager            *podManager
 
 	enableMetrics bool
 	storage       storage
 }
 
 // NewManager returns a new node shutdown manager.
-func NewManager(conf *Config) (Manager, lifecycle.PodAdmitHandler) {
+func NewManager(conf *Config) Manager {
 	if !utilfeature.DefaultFeatureGate.Enabled(features.GracefulNodeShutdown) {
 		m := managerStub{}
-		return m, m
+		return m
 	}
 
-	shutdownGracePeriodByPodPriority := conf.ShutdownGracePeriodByPodPriority
-	// Migration from the original configuration
-	if !utilfeature.DefaultFeatureGate.Enabled(features.GracefulNodeShutdownBasedOnPodPriority) ||
-		len(shutdownGracePeriodByPodPriority) == 0 {
-		shutdownGracePeriodByPodPriority = migrateConfig(conf.ShutdownGracePeriodRequested, conf.ShutdownGracePeriodCriticalPods)
-	}
+	podManager := newPodManager(conf)
 
 	// Disable if the configuration is empty
-	if len(shutdownGracePeriodByPodPriority) == 0 {
+	if len(podManager.shutdownGracePeriodByPodPriority) == 0 {
 		m := managerStub{}
-		return m, m
+		return m
 	}
 
-	// Sort by priority from low to high
-	sort.Slice(shutdownGracePeriodByPodPriority, func(i, j int) bool {
-		return shutdownGracePeriodByPodPriority[i].Priority < shutdownGracePeriodByPodPriority[j].Priority
-	})
-
-	if conf.Clock == nil {
-		conf.Clock = clock.RealClock{}
-	}
 	manager := &managerImpl{
-		logger:                           conf.Logger,
-		probeManager:                     conf.ProbeManager,
-		recorder:                         conf.Recorder,
-		volumeManager:                    conf.VolumeManager,
-		nodeRef:                          conf.NodeRef,
-		getPods:                          conf.GetPodsFunc,
-		killPodFunc:                      conf.KillPodFunc,
-		syncNodeStatus:                   conf.SyncNodeStatusFunc,
-		shutdownGracePeriodByPodPriority: shutdownGracePeriodByPodPriority,
-		clock:                            conf.Clock,
-		enableMetrics:                    utilfeature.DefaultFeatureGate.Enabled(features.GracefulNodeShutdownBasedOnPodPriority),
+		logger:         conf.Logger,
+		probeManager:   conf.ProbeManager,
+		recorder:       conf.Recorder,
+		nodeRef:        conf.NodeRef,
+		getPods:        conf.GetPodsFunc,
+		syncNodeStatus: conf.SyncNodeStatusFunc,
+		podManager:     podManager,
+		enableMetrics:  utilfeature.DefaultFeatureGate.Enabled(features.GracefulNodeShutdownBasedOnPodPriority),
 		storage: localStorage{
 			Path: filepath.Join(conf.StateDirectory, localStorageStateFile),
 		},
@@ -142,9 +111,9 @@ func NewManager(conf *Config) (Manager, lifecycle.PodAdmitHandler) {
 	manager.logger.Info("Creating node shutdown manager",
 		"shutdownGracePeriodRequested", conf.ShutdownGracePeriodRequested,
 		"shutdownGracePeriodCriticalPods", conf.ShutdownGracePeriodCriticalPods,
-		"shutdownGracePeriodByPodPriority", shutdownGracePeriodByPodPriority,
+		"shutdownGracePeriodByPodPriority", podManager.shutdownGracePeriodByPodPriority,
 	)
-	return manager, manager
+	return manager
 }
 
 // Admit rejects all pods if node is shutting
@@ -217,7 +186,7 @@ func (m *managerImpl) start() (chan struct{}, error) {
 	}
 
 	// If the logind's InhibitDelayMaxUSec as configured in (logind.conf) is less than periodRequested, attempt to update the value to periodRequested.
-	if periodRequested := m.periodRequested(); periodRequested > currentInhibitDelay {
+	if periodRequested := m.podManager.periodRequested(); periodRequested > currentInhibitDelay {
 		err := m.dbusCon.OverrideInhibitDelay(periodRequested)
 		if err != nil {
 			return nil, fmt.Errorf("unable to override inhibit delay by shutdown manager: %v", err)
@@ -356,166 +325,5 @@ func (m *managerImpl) processShutdownEvent() error {
 		}()
 	}
 
-	groups := groupByPriority(m.shutdownGracePeriodByPodPriority, activePods)
-	for _, group := range groups {
-		// If there are no pods in a particular range,
-		// then do not wait for pods in that priority range.
-		if len(group.Pods) == 0 {
-			continue
-		}
-
-		var wg sync.WaitGroup
-		wg.Add(len(group.Pods))
-		for _, pod := range group.Pods {
-			go func(pod *v1.Pod, group podShutdownGroup) {
-				defer wg.Done()
-
-				gracePeriodOverride := group.ShutdownGracePeriodSeconds
-
-				// If the pod's spec specifies a termination gracePeriod which is less than the gracePeriodOverride calculated, use the pod spec termination gracePeriod.
-				if pod.Spec.TerminationGracePeriodSeconds != nil && *pod.Spec.TerminationGracePeriodSeconds <= gracePeriodOverride {
-					gracePeriodOverride = *pod.Spec.TerminationGracePeriodSeconds
-				}
-
-				m.logger.V(1).Info("Shutdown manager killing pod with gracePeriod", "pod", klog.KObj(pod), "gracePeriod", gracePeriodOverride)
-
-				if err := m.killPodFunc(pod, false, &gracePeriodOverride, func(status *v1.PodStatus) {
-					// set the pod status to failed (unless it was already in a successful terminal phase)
-					if status.Phase != v1.PodSucceeded {
-						status.Phase = v1.PodFailed
-					}
-					status.Message = nodeShutdownMessage
-					status.Reason = nodeShutdownReason
-					podutil.UpdatePodCondition(status, &v1.PodCondition{
-						Type:    v1.DisruptionTarget,
-						Status:  v1.ConditionTrue,
-						Reason:  v1.PodReasonTerminationByKubelet,
-						Message: nodeShutdownMessage,
-					})
-				}); err != nil {
-					m.logger.V(1).Info("Shutdown manager failed killing pod", "pod", klog.KObj(pod), "err", err)
-				} else {
-					m.logger.V(1).Info("Shutdown manager finished killing pod", "pod", klog.KObj(pod))
-				}
-			}(pod, group)
-		}
-
-		// This duration determines how long the shutdown manager will wait for the pods in this group
-		// to terminate before proceeding to the next group.
-		var groupTerminationWaitDuration = time.Duration(group.ShutdownGracePeriodSeconds) * time.Second
-		var (
-			doneCh         = make(chan struct{})
-			timer          = m.clock.NewTimer(groupTerminationWaitDuration)
-			ctx, ctxCancel = context.WithTimeout(context.Background(), groupTerminationWaitDuration)
-		)
-		go func() {
-			defer close(doneCh)
-			defer ctxCancel()
-			wg.Wait()
-			// The signal to kill a Pod was sent successfully to all the pods,
-			// let's wait until all the volumes are unmounted from all the pods before
-			// continuing to the next group. This is done so that the CSI Driver (assuming
-			// that it's part of the highest group) has a chance to perform unmounts.
-			if err := m.volumeManager.WaitForAllPodsUnmount(ctx, group.Pods); err != nil {
-				var podIdentifiers []string
-				for _, pod := range group.Pods {
-					podIdentifiers = append(podIdentifiers, fmt.Sprintf("%s/%s", pod.Namespace, pod.Name))
-				}
-
-				// Waiting for volume teardown is done on a best basis effort,
-				// report an error and continue.
-				//
-				// Depending on the user provided kubelet configuration value
-				// either the `timer` will tick and we'll continue to shutdown the next group, or,
-				// WaitForAllPodsUnmount will timeout, therefore this goroutine
-				// will close doneCh and we'll continue to shutdown the next group.
-				m.logger.Error(err, "Failed while waiting for all the volumes belonging to Pods in this group to unmount", "pods", podIdentifiers)
-			}
-		}()
-
-		select {
-		case <-doneCh:
-			timer.Stop()
-		case <-timer.C():
-			ctxCancel()
-			m.logger.V(1).Info("Shutdown manager pod killing time out", "gracePeriod", group.ShutdownGracePeriodSeconds, "priority", group.Priority)
-		}
-	}
-
-	return nil
-}
-
-func (m *managerImpl) periodRequested() time.Duration {
-	var sum int64
-	for _, period := range m.shutdownGracePeriodByPodPriority {
-		sum += period.ShutdownGracePeriodSeconds
-	}
-	return time.Duration(sum) * time.Second
-}
-
-func migrateConfig(shutdownGracePeriodRequested, shutdownGracePeriodCriticalPods time.Duration) []kubeletconfig.ShutdownGracePeriodByPodPriority {
-	if shutdownGracePeriodRequested == 0 {
-		return nil
-	}
-	defaultPriority := shutdownGracePeriodRequested - shutdownGracePeriodCriticalPods
-	if defaultPriority < 0 {
-		return nil
-	}
-	criticalPriority := shutdownGracePeriodRequested - defaultPriority
-	if criticalPriority < 0 {
-		return nil
-	}
-	return []kubeletconfig.ShutdownGracePeriodByPodPriority{
-		{
-			Priority:                   scheduling.DefaultPriorityWhenNoDefaultClassExists,
-			ShutdownGracePeriodSeconds: int64(defaultPriority / time.Second),
-		},
-		{
-			Priority:                   scheduling.SystemCriticalPriority,
-			ShutdownGracePeriodSeconds: int64(criticalPriority / time.Second),
-		},
-	}
-}
-
-func groupByPriority(shutdownGracePeriodByPodPriority []kubeletconfig.ShutdownGracePeriodByPodPriority, pods []*v1.Pod) []podShutdownGroup {
-	groups := make([]podShutdownGroup, 0, len(shutdownGracePeriodByPodPriority))
-	for _, period := range shutdownGracePeriodByPodPriority {
-		groups = append(groups, podShutdownGroup{
-			ShutdownGracePeriodByPodPriority: period,
-		})
-	}
-
-	for _, pod := range pods {
-		var priority int32
-		if pod.Spec.Priority != nil {
-			priority = *pod.Spec.Priority
-		}
-
-		// Find the group index according to the priority.
-		index := sort.Search(len(groups), func(i int) bool {
-			return groups[i].Priority >= priority
-		})
-
-		// 1. Those higher than the highest priority default to the highest priority
-		// 2. Those lower than the lowest priority default to the lowest priority
-		// 3. Those boundary priority default to the lower priority
-		// if priority of pod is:
-		//   groups[index-1].Priority <= pod priority < groups[index].Priority
-		// in which case we want to pick lower one (i.e index-1)
-		if index == len(groups) {
-			index = len(groups) - 1
-		} else if index < 0 {
-			index = 0
-		} else if index > 0 && groups[index].Priority > priority {
-			index--
-		}
-
-		groups[index].Pods = append(groups[index].Pods, pod)
-	}
-	return groups
-}
-
-type podShutdownGroup struct {
-	kubeletconfig.ShutdownGracePeriodByPodPriority
-	Pods []*v1.Pod
+	return m.podManager.killPods(activePods)
 }
