+++ pkg/kubelet/nodeshutdown/nodeshutdown_manager_linux_test.go
@@ -92,19 +92,6 @@ func (f *fakeDbus) OverrideInhibitDelay(inhibitDelayMax time.Duration) error {
 	return nil
 }
 
-func makePod(name string, priority int32, terminationGracePeriod *int64) *v1.Pod {
-	return &v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-			UID:  types.UID(name),
-		},
-		Spec: v1.PodSpec{
-			Priority:                      &priority,
-			TerminationGracePeriodSeconds: terminationGracePeriod,
-		},
-	}
-}
-
 func TestManager(t *testing.T) {
 	systemDbusTmp := systemDbus
 	defer func() {
@@ -352,7 +339,7 @@ func TestManager(t *testing.T) {
 			fakeRecorder := &record.FakeRecorder{}
 			fakeVolumeManager := volumemanager.NewFakeVolumeManager([]v1.UniqueVolumeName{}, 0, nil)
 			nodeRef := &v1.ObjectReference{Kind: "Node", Name: "test", UID: types.UID("test"), Namespace: ""}
-			manager, _ := NewManager(&Config{
+			manager := NewManager(&Config{
 				Logger:                          logger,
 				ProbeManager:                    proberManager,
 				VolumeManager:                   fakeVolumeManager,
@@ -459,7 +446,7 @@ func TestFeatureEnabled(t *testing.T) {
 			fakeVolumeManager := volumemanager.NewFakeVolumeManager([]v1.UniqueVolumeName{}, 0, nil)
 			nodeRef := &v1.ObjectReference{Kind: "Node", Name: "test", UID: types.UID("test"), Namespace: ""}
 
-			manager, _ := NewManager(&Config{
+			manager := NewManager(&Config{
 				Logger:                          logger,
 				ProbeManager:                    proberManager,
 				VolumeManager:                   fakeVolumeManager,
@@ -517,7 +504,7 @@ func TestRestart(t *testing.T) {
 	fakeRecorder := &record.FakeRecorder{}
 	fakeVolumeManager := volumemanager.NewFakeVolumeManager([]v1.UniqueVolumeName{}, 0, nil)
 	nodeRef := &v1.ObjectReference{Kind: "Node", Name: "test", UID: types.UID("test"), Namespace: ""}
-	manager, _ := NewManager(&Config{
+	manager := NewManager(&Config{
 		Logger:                          logger,
 		ProbeManager:                    proberManager,
 		VolumeManager:                   fakeVolumeManager,
@@ -548,199 +535,6 @@ func TestRestart(t *testing.T) {
 		shutdownChanMut.Lock()
 		close(shutdownChan)
 		shutdownChanMut.Unlock()
-	}
-}
-
-func Test_migrateConfig(t *testing.T) {
-	type shutdownConfig struct {
-		shutdownGracePeriodRequested    time.Duration
-		shutdownGracePeriodCriticalPods time.Duration
-	}
-	tests := []struct {
-		name string
-		args shutdownConfig
-		want []kubeletconfig.ShutdownGracePeriodByPodPriority
-	}{
-		{
-			name: "both shutdownGracePeriodRequested and shutdownGracePeriodCriticalPods",
-			args: shutdownConfig{
-				shutdownGracePeriodRequested:    300 * time.Second,
-				shutdownGracePeriodCriticalPods: 120 * time.Second,
-			},
-			want: []kubeletconfig.ShutdownGracePeriodByPodPriority{
-				{
-					Priority:                   scheduling.DefaultPriorityWhenNoDefaultClassExists,
-					ShutdownGracePeriodSeconds: 180,
-				},
-				{
-					Priority:                   scheduling.SystemCriticalPriority,
-					ShutdownGracePeriodSeconds: 120,
-				},
-			},
-		},
-		{
-			name: "only shutdownGracePeriodRequested",
-			args: shutdownConfig{
-				shutdownGracePeriodRequested:    100 * time.Second,
-				shutdownGracePeriodCriticalPods: 0 * time.Second,
-			},
-			want: []kubeletconfig.ShutdownGracePeriodByPodPriority{
-				{
-					Priority:                   scheduling.DefaultPriorityWhenNoDefaultClassExists,
-					ShutdownGracePeriodSeconds: 100,
-				},
-				{
-					Priority:                   scheduling.SystemCriticalPriority,
-					ShutdownGracePeriodSeconds: 0,
-				},
-			},
-		},
-		{
-			name: "empty configuration",
-			args: shutdownConfig{
-				shutdownGracePeriodRequested:    0 * time.Second,
-				shutdownGracePeriodCriticalPods: 0 * time.Second,
-			},
-			want: nil,
-		},
-		{
-			name: "wrong configuration",
-			args: shutdownConfig{
-				shutdownGracePeriodRequested:    1 * time.Second,
-				shutdownGracePeriodCriticalPods: 100 * time.Second,
-			},
-			want: nil,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := migrateConfig(tt.args.shutdownGracePeriodRequested, tt.args.shutdownGracePeriodCriticalPods); !assert.Equal(t, tt.want, got) {
-				t.Errorf("migrateConfig() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_groupByPriority(t *testing.T) {
-	type args struct {
-		shutdownGracePeriodByPodPriority []kubeletconfig.ShutdownGracePeriodByPodPriority
-		pods                             []*v1.Pod
-	}
-	tests := []struct {
-		name string
-		args args
-		want []podShutdownGroup
-	}{
-		{
-			name: "migrate config",
-			args: args{
-				shutdownGracePeriodByPodPriority: migrateConfig(300*time.Second /* shutdownGracePeriodRequested */, 120*time.Second /* shutdownGracePeriodCriticalPods */),
-				pods: []*v1.Pod{
-					makePod("normal-pod", scheduling.DefaultPriorityWhenNoDefaultClassExists, nil),
-					makePod("highest-user-definable-pod", scheduling.HighestUserDefinablePriority, nil),
-					makePod("critical-pod", scheduling.SystemCriticalPriority, nil),
-				},
-			},
-			want: []podShutdownGroup{
-				{
-					ShutdownGracePeriodByPodPriority: kubeletconfig.ShutdownGracePeriodByPodPriority{
-						Priority:                   scheduling.DefaultPriorityWhenNoDefaultClassExists,
-						ShutdownGracePeriodSeconds: 180,
-					},
-					Pods: []*v1.Pod{
-						makePod("normal-pod", scheduling.DefaultPriorityWhenNoDefaultClassExists, nil),
-						makePod("highest-user-definable-pod", scheduling.HighestUserDefinablePriority, nil),
-					},
-				},
-				{
-					ShutdownGracePeriodByPodPriority: kubeletconfig.ShutdownGracePeriodByPodPriority{
-						Priority:                   scheduling.SystemCriticalPriority,
-						ShutdownGracePeriodSeconds: 120,
-					},
-					Pods: []*v1.Pod{
-						makePod("critical-pod", scheduling.SystemCriticalPriority, nil),
-					},
-				},
-			},
-		},
-		{
-			name: "pod priority",
-			args: args{
-				shutdownGracePeriodByPodPriority: []kubeletconfig.ShutdownGracePeriodByPodPriority{
-					{
-						Priority:                   1,
-						ShutdownGracePeriodSeconds: 10,
-					},
-					{
-						Priority:                   2,
-						ShutdownGracePeriodSeconds: 20,
-					},
-					{
-						Priority:                   3,
-						ShutdownGracePeriodSeconds: 30,
-					},
-					{
-						Priority:                   4,
-						ShutdownGracePeriodSeconds: 40,
-					},
-				},
-				pods: []*v1.Pod{
-					makePod("pod-0", 0, nil),
-					makePod("pod-1", 1, nil),
-					makePod("pod-2", 2, nil),
-					makePod("pod-3", 3, nil),
-					makePod("pod-4", 4, nil),
-					makePod("pod-5", 5, nil),
-				},
-			},
-			want: []podShutdownGroup{
-				{
-					ShutdownGracePeriodByPodPriority: kubeletconfig.ShutdownGracePeriodByPodPriority{
-						Priority:                   1,
-						ShutdownGracePeriodSeconds: 10,
-					},
-					Pods: []*v1.Pod{
-						makePod("pod-0", 0, nil),
-						makePod("pod-1", 1, nil),
-					},
-				},
-				{
-					ShutdownGracePeriodByPodPriority: kubeletconfig.ShutdownGracePeriodByPodPriority{
-						Priority:                   2,
-						ShutdownGracePeriodSeconds: 20,
-					},
-					Pods: []*v1.Pod{
-						makePod("pod-2", 2, nil),
-					},
-				},
-				{
-					ShutdownGracePeriodByPodPriority: kubeletconfig.ShutdownGracePeriodByPodPriority{
-						Priority:                   3,
-						ShutdownGracePeriodSeconds: 30,
-					},
-					Pods: []*v1.Pod{
-						makePod("pod-3", 3, nil),
-					},
-				},
-				{
-					ShutdownGracePeriodByPodPriority: kubeletconfig.ShutdownGracePeriodByPodPriority{
-						Priority:                   4,
-						ShutdownGracePeriodSeconds: 40,
-					},
-					Pods: []*v1.Pod{
-						makePod("pod-4", 4, nil),
-						makePod("pod-5", 5, nil),
-					},
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := groupByPriority(tt.args.shutdownGracePeriodByPodPriority, tt.args.pods); !assert.Equal(t, tt.want, got) {
-				t.Errorf("groupByPriority() = %v, want %v", got, tt.want)
-			}
-		})
 	}
 }
 
@@ -818,20 +612,23 @@ func Test_managerImpl_processShutdownEvent(t *testing.T) {
 				),
 			)
 			m := &managerImpl{
-				logger:                           logger,
-				volumeManager:                    tt.fields.volumeManager,
-				recorder:                         tt.fields.recorder,
-				nodeRef:                          tt.fields.nodeRef,
-				probeManager:                     tt.fields.probeManager,
-				shutdownGracePeriodByPodPriority: tt.fields.shutdownGracePeriodByPodPriority,
-				getPods:                          tt.fields.getPods,
-				killPodFunc:                      tt.fields.killPodFunc,
-				syncNodeStatus:                   tt.fields.syncNodeStatus,
-				dbusCon:                          tt.fields.dbusCon,
-				inhibitLock:                      tt.fields.inhibitLock,
-				nodeShuttingDownMutex:            sync.Mutex{},
-				nodeShuttingDownNow:              tt.fields.nodeShuttingDownNow,
-				clock:                            tt.fields.clock,
+				logger:                logger,
+				recorder:              tt.fields.recorder,
+				nodeRef:               tt.fields.nodeRef,
+				probeManager:          tt.fields.probeManager,
+				getPods:               tt.fields.getPods,
+				syncNodeStatus:        tt.fields.syncNodeStatus,
+				dbusCon:               tt.fields.dbusCon,
+				inhibitLock:           tt.fields.inhibitLock,
+				nodeShuttingDownMutex: sync.Mutex{},
+				nodeShuttingDownNow:   tt.fields.nodeShuttingDownNow,
+				podManager: &podManager{
+					logger:                           logger,
+					volumeManager:                    tt.fields.volumeManager,
+					shutdownGracePeriodByPodPriority: tt.fields.shutdownGracePeriodByPodPriority,
+					killPodFunc:                      tt.fields.killPodFunc,
+					clock:                            tt.fields.clock,
+				},
 			}
 			if err := m.processShutdownEvent(); (err != nil) != tt.wantErr {
 				t.Errorf("managerImpl.processShutdownEvent() error = %v, wantErr %v", err, tt.wantErr)
@@ -870,28 +667,31 @@ func Test_processShutdownEvent_VolumeUnmountTimeout(t *testing.T) {
 	)
 	logger := ktesting.NewLogger(t, ktesting.NewConfig(ktesting.BufferLogs(true)))
 	m := &managerImpl{
-		logger:        logger,
-		volumeManager: fakeVolumeManager,
-		recorder:      fakeRecorder,
-		nodeRef:       nodeRef,
-		probeManager:  probeManager,
-		shutdownGracePeriodByPodPriority: []kubeletconfig.ShutdownGracePeriodByPodPriority{
-			{
-				Priority:                   1,
-				ShutdownGracePeriodSeconds: int64(shutdownGracePeriodSeconds),
-			},
-		},
+		logger:       logger,
+		recorder:     fakeRecorder,
+		nodeRef:      nodeRef,
+		probeManager: probeManager,
 		getPods: func() []*v1.Pod {
 			return []*v1.Pod{
 				makePod("test-pod", 1, nil),
 			}
 		},
-		killPodFunc: func(pod *v1.Pod, isEvicted bool, gracePeriodOverride *int64, fn func(*v1.PodStatus)) error {
-			return nil
-		},
 		syncNodeStatus: syncNodeStatus,
 		dbusCon:        &fakeDbus{},
-		clock:          fakeclock,
+		podManager: &podManager{
+			logger:        logger,
+			volumeManager: fakeVolumeManager,
+			shutdownGracePeriodByPodPriority: []kubeletconfig.ShutdownGracePeriodByPodPriority{
+				{
+					Priority:                   1,
+					ShutdownGracePeriodSeconds: int64(shutdownGracePeriodSeconds),
+				},
+			},
+			killPodFunc: func(pod *v1.Pod, isEvicted bool, gracePeriodOverride *int64, fn func(*v1.PodStatus)) error {
+				return nil
+			},
+			clock: fakeclock,
+		},
 	}
 
 	start := fakeclock.Now()
